isolate proxy/firewall config from quick start generated config for easy upgrades

+++ remove.sh
@@ -26,8 +26,8 @@ echo "${bold}*************************************"
 echo "Quorum Dev Quickstart "
 echo "*************************************${normal}"
 echo "Stop and remove network..."
-docker-compose down -v
-docker-compose rm -sfv
+docker-compose -f docker-compose.yml -f docker-compose-override.yml down -v
+docker-compose -f docker-compose.yml -f docker-compose-override.yml rm -sfv
 
 if [ -f "docker-compose-deps.yml" ]; then
     echo "Stopping dependencies..."
@@ -41,11 +41,35 @@ if [[ ! -z `docker ps -a | grep quorum-dev-quickstart_pet_shop` ]]; then
   removeDockerImage quorum-dev-quickstart_pet_shop
 fi
 
-if grep -q 'kibana:' docker-compose.yml 2> /dev/null ; then
-  docker image rm quorum-test-network_elasticsearch
-  docker image rm quorum-test-network_logstash
-  docker image rm quorum-test-network_filebeat
-  docker image rm quorum-test-network_metricbeat
+rm_image=${1:-nrmi}
+
+if [ $rm_image == "-rmi" ]; then
+  echo "Removing images"
+  docker image rm quorum-dev-quickstart/block-explorer-light:develop
+  docker image rm consensys/quorum-ethsigner:${QUORUM_ETHSIGNER_VERSION}
+  if grep -q 'orion:' docker-compose.yml 2> /dev/null ; then
+    docker image rm consensys/quorum-orion:${QUORUM_ORION_VERSION}
+  fi
+  if grep -q 'besu:' docker-compose.yml 2> /dev/null ; then
+    docker image rm hyperledger/besu:${BESU_VERSION}
+  fi
+  if grep -q 'tessera:' docker-compose.yml 2> /dev/null ; then
+    docker image rm quorumengineering/tessera:${QUORUM_TESSERA_VERSION}
+  fi
+  if grep -q 'quorum:' docker-compose.yml 2> /dev/null ; then
+    docker image rm quorumengineering/quorum:${QUORUM_VERSION}
+  fi
+  if grep -q 'cakeshop:' docker-compose.yml 2> /dev/null ; then
+    docker image rm quorumengineering/cakeshop:${QUORUM_CAKESHOP_VERSION}
+  fi
+  if grep -q 'kibana:' docker-compose.yml 2> /dev/null ; then
+    docker image rm quorum-test-network_elasticsearch
+    docker image rm quorum-test-network_logstash
+    docker image rm quorum-test-network_filebeat
+    docker image rm quorum-test-network_metricbeat
+  fi
+else
+  echo "Skipping image removal"
 fi
 
 rm ${LOCK_FILE}
